fix: Fixed event handling for multiple sdl windows.
Created a new function that collects all SDL events into a local buffer
that is then iterated over to find the first event matching the given
window id. If one is found it's returned in the out parameter and the
function succeeds. If none is found then it returns false.