chore: Trying to fix the example app conan setup.
Instead of adding the subdirectory in cmake, the example app now
properly requires the conan package canyon and then links to it in
cmake.
This still fails however because dependencies in canyon or moth, like
vulkan and nlohmann json, need to be added to the example app. This
seems wrong to me so I'm looking into how to fix it.