fix: ITargets are no longer IImages. Fixes image rendering.
ITargets being IImages broke some assumptions in the image rendering.
The passed in IImage was assumed to always be an Image type from vulkan,
but sometimes it was a FrameBuffer.
This has been fixed by no longer treating ITargets as IImages but they
now have a method to fetch the underlying IImage which is the correct
type.